Add Nodesax changes
