+++ src/main/java/com/totalcross/knowcode/ui/CustomInitUI.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-3.0-only
   *********************************************************************************/
 
-package com.totalcross.knowcode.parse;
+package com.totalcross.knowcode.ui;
 
 /**
  * CustomInitUI is a interface to edit the components of the XML file before swap the window.
