+++ src/main/java/com/totalcross/knowcode/xml/NodeSaxAbsolute.java
@@ -1,0 +1,17 @@
+package com.totalcross.knowcode.xml;
+
+import totalcross.io.IOException;
+import totalcross.ui.image.ImageException;
+
+public class NodeSaxAbsolute extends NodeSax{
+
+    /**
+     * Creates a NodeSaxConstraintAbsolute object
+     *
+     * @throws ImageException
+     * @throws IOException
+     */
+    public NodeSaxAbsolute(NodeSax node) throws IOException, ImageException {
+        super(node);
+    }
+}
