+++ src/main/java/com/totalcross/knowcode/xml/NodeSaxRelative.java
@@ -1,0 +1,21 @@
+package com.totalcross.knowcode.xml;
+
+import totalcross.io.IOException;
+
+import totalcross.ui.image.ImageException;
+
+
+public class NodeSaxRelative extends NodeSax {
+	
+    /**
+     * Creates a NodeSaxRelative object
+     *
+     * @throws ImageException
+     * @throws IOException
+     */
+    public NodeSaxRelative(NodeSax node) throws IOException, ImageException {
+        super(node);
+    }
+
+
+}
